Do not gevent.monkey.patch stuff so it can be imported by others.
To use greenlets, gevent needs to monkey patch a bunch of stuff.  This
needs to be done as early possible and in the main thread.  By doing
it in our own modules, we prevents imsearchtools from being used by
other programs, which get warnings about monkey patching too late or
being unable to switch thread.

This commit removes the monkey patching.  If some programs wants to
make use of imsearchtools, then they need to call gevent.monkey_*
themselves when it suits them.