ImageGetter: save images from content instead of raw.
response.raw is a raw stream of bytes – it does not transform the
response content, namely it does not decode the gzip and deflate
transfer-encodings.  When we receive images in gzip encoded we then we
get a "cannot identify image file" error.  So save the content.

Also, since we get the images with stream=True, use iter_content.