Replace hardcoded API keys with individual arguments to SearchClient class.
Most search engines require some sort of API key.  These need to be
hardcoded into the imsearchtools.engines.api_credentials module.  This
is a pain to redistribute, one needs to edit the source and rebuild.

This replaces that approach with requiring the API keys in the
individual SearchClient classes.  http_service has new optional
arguments for those keys which causes the corresponding engines to be
enabled or disabled.  The api_credentials module is removed since it
is no longer used and its documentation moved to the individual engine
modules.